Initial commit of (broken) CUDA code.
I wrote this hours ago, and have been spending my time dealing with
how broken OpenCV is when it comes to CUDA.  It's (understandably)
disabled in the Debian package, so I build my own.  It takes a couple
of hours.

Then I find out that the Haar cascade classifiers are disabled (in
source, not configuration) for no good reason.  Fix.  The CUDA MJPEG
decoder is disabled (even though the bug that was the reason has since
been fixed); fix.  OpenGL is disabled by default in OpenCV; why?!?

Now I need to get back to the working (CPU-based) version, so I'm
putting this on a branch.  I'll get back to it maybe after my build
finishes.